+++ src/main/java/com/trustamarket/inspectionservice/center/domain/enums/CenterStatus.java
@@ -1,0 +1,7 @@
+package com.trustamarket.inspectionservice.center.domain.enums;
+
+public enum CenterStatus {
+    OPEN,
+    MAINTENANCE,
+    CLOSED
+}
