+++ src/main/java/com/trustamarket/inspectionservice/center/domain/exception/InspectionCenterException.java
@@ -1,8 +1,6 @@
-package com.trustamarket.inspectionservice.center.domain.model.exception;
+package com.trustamarket.inspectionservice.center.domain.exception;
 
-// TODO: 추후에 common 내의 BassException 상속 구조로 변경
 public class InspectionCenterException extends RuntimeException {
-
     public InspectionCenterException(String message) {
         super(message);
     }
