+++ src/main/java/com/trustamarket/inspectionservice/center/domain/model/InspectionCenter.java
@@ -1,9 +1,9 @@
 package com.trustamarket.inspectionservice.center.domain.model;
 
-import com.trustamarket.inspectionservice.center.domain.model.enums.CenterStatus;
-import com.trustamarket.inspectionservice.center.domain.model.exception.InspectionCenterException;
-import com.trustamarket.inspectionservice.center.domain.model.vo.Address;
-import com.trustamarket.inspectionservice.center.domain.model.vo.CenterId;
+import com.trustamarket.inspectionservice.center.domain.enums.CenterStatus;
+import com.trustamarket.inspectionservice.center.domain.exception.InspectionCenterException;
+import com.trustamarket.inspectionservice.center.domain.vo.Address;
+import com.trustamarket.inspectionservice.center.domain.vo.CenterId;
 import lombok.Getter;
 
 import java.util.Objects;
