+++ src/main/java/com/trustamarket/inspectionservice/inspection/domain/enums/CurrencyCode.java
@@ -1,0 +1,5 @@
+package com.trustamarket.inspectionservice.inspection.domain.enums;
+
+public enum CurrencyCode {
+    KRW
+}
