+++ src/main/java/com/trustamarket/inspectionservice/inspection/domain/enums/InspectionType.java
@@ -1,0 +1,6 @@
+package com.trustamarket.inspectionservice.inspection.domain.enums;
+
+public enum InspectionType {
+    INITIAL,
+    RETURN
+}
