+++ src/main/java/com/trustamarket/inspectionservice/inspection/domain/enums/PhotoType.java
@@ -1,4 +1,4 @@
-package com.trustamarket.inspectionservice.inspection.domain.model.enums;
+package com.trustamarket.inspectionservice.inspection.domain.enums;
 
 public enum PhotoType {
     EXTERIOR,
