--- src/main/java/com/trustamarket/inspectionservice/inspection/domain/model/enums/Grade.java
@@ -1,5 +1,0 @@
-package com.trustamarket.inspectionservice.inspection.domain.model.enums;
-
-public enum Grade {
-    S, A, B, C
-}
