--- src/main/java/com/trustamarket/inspectionservice/inspection/domain/model/enums/InspectionType.java
@@ -1,6 +1,0 @@
-package com.trustamarket.inspectionservice.inspection.domain.model.enums;
-
-public enum InspectionType {
-    INITIAL,
-    RETURN
-}
