+++ src/main/java/com/trustamarket/inspectionservice/inspection/domain/vo/SellerId.java
@@ -1,10 +1,9 @@
-package com.trustamarket.inspectionservice.inspection.domain.model.vo;
+package com.trustamarket.inspectionservice.inspection.domain.vo;
 
 import java.util.Objects;
 import java.util.UUID;
 
 public record SellerId(UUID value) {
-
     public SellerId {
         Objects.requireNonNull(value, "SellerId 값은 필수입니다");
     }
